更新sdk，新增bridge、require统计；新增统计分析工具

+++ MLN-Android/mlnservics/src/main/java/com/immomo/mls/fun/other/Size.java
@@ -40,7 +40,7 @@ public class Size {
     }
 
     public float getWidth() {
-        return (int) width;
+        return width;
     }
 
     public void setWidth(float width) {
@@ -48,7 +48,7 @@ public class Size {
     }
 
     public float getHeight() {
-        return (int) height;
+        return height;
     }
 
     public void setHeight(float height) {
